totalStudyMinutesの型をLongに変更 ・SUM() の結果は Long になるため、データによっては変換エラーが発生する可能性があるため

+++ src/main/java/com/example/studytracker/dto/calendar/CalendarResponse.java
@@ -25,5 +25,5 @@ public class CalendarResponse {
     /**
      * 学習合計時間（分）
      */
-    private Integer totalStudyMinutes;
+    private Long totalStudyMinutes;
 }
