+++ src/main/java/com/example/studytracker/dto/stat/MonthlyStatsResponse.java
@@ -23,5 +23,5 @@ public class MonthlyStatsResponse {
     /**
      * 学習合計時間（分）
      */
-    private Integer totalStudyMinutes;
+    private Long totalStudyMinutes;
 }
